% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{age_to_time}
\alias{age_to_time}
\alias{age_to_time.surv_table}
\alias{age_to_time.surv_object}
\alias{age_to_time.surv_pooled}
\title{Convert an age-based table to a cycle-based table}
\usage{
age_to_time(x, age_init, cycle_length)

\method{age_to_time}{surv_table}(x, age_init, cycle_length)

\method{age_to_time}{surv_object}(x, age_init, cycle_length)

\method{age_to_time}{surv_pooled}(x, age_init, cycle_length)
}
\arguments{
\item{x}{a survival object}

\item{age_init}{initial age the table will be applied to}

\item{cycle_length}{length of cycles}
}
\value{
a table with \code{time} renormalized to be in terms of cycles,
with time 0 corresponding to age \code{age_init}
}
\description{
Convert an age-based table to a cycle-based table
}
\details{
This is used to convert a mortality table for use as a
survival curve in terms of cycles of the model.  Typically,
age_surv will have been constructed using \code{\link[=define_surv_table]{define_surv_table()}}
or \code{\link[=define_wtd_surv_table]{define_wtd_surv_table()}}.
}
\examples{

 df <- data.frame(time = c(50, 51, 55, 60), survival = c(1, 0.9, 0.7, 0.5))
 st <- define_surv_table(df, by_age = TRUE)
 age_to_time(st, age_init = 50, cycle_length = 1)
}
