% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_general.R
\name{build_pieces_def_structure}
\alias{build_pieces_def_structure}
\title{Build up general piecewise distribution specification
from the component pieces}
\usage{
build_pieces_def_structure(dists, breaks, params)
}
\arguments{
\item{dists}{Character vector naming the distributions for each piece.}

\item{breaks}{Breakpoints at which distributions change.}

\item{params}{parameters of the hazard functions for the distributions.}
}
\value{
A data frame with columns \code{breaks}, \code{dist_name},
and one additional column for each argument (other than \code{x}
or \code{log}) of the arguments to the hazard functions for the
distributions.
}
\description{
Build up general piecewise distribution specification
from the component pieces
}
\examples{
\dontrun{
## requires package flexsurv to be loaded
build_pieces_def_structure(c("exp", "weibull", "exp"),
                           c(0, 100, 200),
                           c(0.005, 1, 100, 0.002))
                           }
}
