% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{compute_counts.eval_part_surv}
\alias{compute_counts.eval_part_surv}
\title{Compute Count of Individual in Each State per Cycle}
\usage{
compute_counts.eval_part_surv(x, init, inflow, ...)
}
\arguments{
\item{x}{An \code{eval_part_surv} object.}

\item{init}{numeric vector, same length as number of
model states. Number of individuals in each model state
at the beginning.}

\item{inflow}{numeric vector, similar to \code{init}.
Number of new individuals in each state per cycle.}

\item{...}{additional arguments (currently ignored).}
}
\value{
A \code{cycle_counts} object.
}
\description{
Compute Count of Individual in Each State per Cycle
}
