% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{cost_at_right_time}
\alias{cost_at_right_time}
\title{Allow assignment of a cost at a particular cycle or age}
\usage{
cost_at_right_time(cost, age, cycle, at_age, at_cycle)
}
\arguments{
\item{cost}{cost, in units when incurred}

\item{age}{age at a given cycle}

\item{cycle}{the cycle}

\item{at_age}{at what age should the cost be incurred?}

\item{at_cycle}{in what cycle should cost be incurred?}
}
\value{
0 except when the age or cycle condition is met;
\code{cost} when at least one condition is met.
}
\description{
Allow assignment of a cost at a particular cycle or age
}
\details{
intended for use in parameters object
for \code{heemod} models.
}
\examples{
cost_at_right_time(500, age = 1:20, at_age = 10)
cost_at_right_time(500, cycle = 1:20, at_cycle = 10)
}
