% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{cost_iv_administration}
\alias{cost_iv_administration}
\title{Cost of administration for an intravenous treatment}
\usage{
cost_iv_administration(iv_time, cost_first_unit, cost_addl_units,
  prorate_first = FALSE)
}
\arguments{
\item{iv_time}{The infusion time, in units compatible with the costs}

\item{cost_first_unit}{cost for the first time unit}

\item{cost_addl_units}{cost for the second time unit}

\item{prorate_first}{can we charge for less than 1 unit of time?
See details.}
}
\value{
the cost
}
\description{
Cost of administration for an intravenous treatment
}
\details{
Simple formula:  cost for the first time unit (prorated if
total time is less than one unit and \code{prorate_first == TRUE})
plus the prorated cost for additional units.

If \code{prorate_first == FALSE}, then the smallest time that
can be charged is one time unit.   For example, if time = 0.5
and \code{prorate_first == TRUE}, the function will charge for one
half of a time unit.   If time = 0.5 and \code{prorate_first == FALSE},
the function will charge for a full time unit.
}
\examples{
cost_iv_administration(0.5, 100, 20, prorate_first = TRUE) # = 50
cost_iv_administration(1.5, 100, 20, prorate_first = TRUE) # = 110
cost_iv_administration(0.5, 100, 20, prorate_first = FALSE) # = 100
cost_iv_administration(1.5, 100, 20, prorate_first = FALSE) # = 110
}
