% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{cost_iv_compound_administration}
\alias{cost_iv_compound_administration}
\title{Cost of administration for an intravenous treatment}
\usage{
cost_iv_compound_administration(data_table, compound, time_col = "iv_time_hr",
  first_cost_col = "cost_admin_first_hr",
  addl_cost_col = "cost_admin_addl_hr", prorate_first = FALSE, ...)
}
\arguments{
\item{data_table}{a data frame; see details.}

\item{compound}{the name of the compound.}

\item{time_col, first_cost_col, addl_cost_col}{parameter names
in data_table.}

\item{prorate_first}{can we charge for less than 1 unit of time?
See details in \code{\link[=cost_iv_administration]{cost_iv_administration()}}.}

\item{...}{additional arguments, filtering criteria
that will be passed to \code{\link[heemod:look_up]{heemod::look_up()}}.}
}
\value{
the cost
}
\description{
Cost of administration for an intravenous treatment
}
\details{
\code{data_table} must have columns \code{compound}, \code{param},
and \code{value}.   \code{time_col}, \code{first_cost_col} and \code{addl_cost_col}
must be names in the \code{param} column.
The required values are found in \code{data_table}
using \code{\link[heemod:look_up]{heemod::look_up()}}.   Then the values are summed using
\code{\link[=cost_iv_administration]{cost_iv_administration()}}.
}
\examples{
exampleParams <- 
  data.frame(compound = "X", 
  param = c("cost_admin_first_hr", "cost_admin_addl_hr",
            "iv_time_hr"),
  value = c(100, 20, 1.5))
  cost_iv_compound_administration(exampleParams, "X")
}
