% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{eval_transition.part_surv}
\alias{eval_transition.part_surv}
\title{Title}
\usage{
eval_transition.part_surv(x, parameters)
}
\arguments{
\item{x}{partitioned survival (\code{part_surv}) objecct}

\item{parameters}{an \code{eval_parameters} object.}
}
\value{
an \code{eval_part_surv} object.
}
\description{
Title
}
