% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{extract_surv_fit_metrics}
\alias{extract_surv_fit_metrics}
\title{Add fit metrics to a fit tibble}
\usage{
extract_surv_fit_metrics(fit_tib, metric = c("AIC", "BIC", "m2LL"))
}
\arguments{
\item{fit_tib}{a tibble of fits, for example from
\code{\link[=survival_fits_from_tabular]{survival_fits_from_tabular()}}}

\item{metric}{which metrics to add}
}
\value{
a tibble with the added columns
}
\description{
Add fit metrics to a fit tibble
}
\examples{
fit_tibble <- survival_fits_from_tabular(system.file("surv",
                                                            package = "heemodFits"), 
                                  "example_oncSpecs.csv", 
                                  c("ProgressionFree", "Progressive", 
                                  "Terminal", "Death"), 
                                  save_fits = FALSE,
                                  just_load = FALSE)
fit_tibble <- extract_surv_fit_metrics(fit_tibble)
}
