% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{fix_censor_col}
\alias{fix_censor_col}
\title{Recode censoring / event column}
\usage{
fix_censor_col(this_data, this_censor_col, censor_code, event_code, file_name)
}
\arguments{
\item{this_data}{data set}

\item{this_censor_col}{the column to change}

\item{censor_code}{code for censoring}

\item{event_code}{code for events}

\item{file_name}{the name of the file this came from,
in case it's needed for an error message}
}
\value{
the same data set, with the column recoded
}
\description{
Recode censoring / event column
}
