% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_find_best_piecewise_survival_models_multi.R
\name{get_best_fit_across_breakpoints}
\alias{get_best_fit_across_breakpoints}
\title{Get best fit across sets of breakpoints}
\usage{
get_best_fit_across_breakpoints(args, use_dists, these_bp, fixed_bp, survtimes,
  eval_metric)
}
\arguments{
\item{args}{arguments for fitting passed from \code{f_find_best_with_fixed_num_pieces}}

\item{use_dists}{the distributions to be used for the pieces between breakpoints}

\item{these_bp}{sets of breakpoints to be compared across}

\item{fixed_bp}{Should breakpoints be considered fixed (the default) or
can they be optimized over?}

\item{survtimes}{the survival times}

\item{eval_metric}{one of \code{AIC}, \code{BIC}, or \code{m2LL}}
}
\value{
a list with 3 components:
\itemize{
\item{\code{fit}; the fit object from \code{flexsurvreg};}
\item{\code{breaks}; the corresponding breakpoints}; and
\item{\code{params}; the parameters for the distributions in the piecewise fit}
}
}
\description{
Get best fit across sets of breakpoints
}
