% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_exp.R
\name{get_best_pe}
\alias{get_best_pe}
\title{Create piecewise exponential definition}
\usage{
get_best_pe(fixed_bp, used_bp, best_fit)
}
\arguments{
\item{fixed_bp}{Should breakpoints be considered fixed?  (Otherwise optimized over.)}

\item{used_bp}{breakpoints}

\item{best_fit}{the fit corresponding to those breakpoints}
}
\value{
a piecewise exponential specification, a data frame with
columns start (the breakpoints) and hazard.
}
\description{
Create piecewise exponential definition
}
