% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_find_best_piecewise_survival_models_multi.R
\name{get_breakpoints}
\alias{get_breakpoints}
\title{Return a set of breakpoints for fitting a piecewise survival model}
\usage{
get_breakpoints(times, num_cand, check_breaks, include_breaks, min_bp_dist,
  num_breaks, fixed_bp)
}
\arguments{
\item{times}{times from survival data}

\item{num_cand}{number of candidate breakpoints checked (selected at
evenly-spaced percentiles of times in survdata).  This number does
not include any breaks in \code{check_breaks} or \code{include_breaks}.}

\item{check_breaks}{points that should be included in the set of
potential breaks (but may not end up being included).}

\item{include_breaks}{breaks that must be included in any solution.}

\item{min_bp_dist}{The minimum distance between adjacent breakpoints.}

\item{num_breaks}{How many breakpoints should there be in the
piecewise survival fit?}

\item{fixed_bp}{Should breakpoints be considered fixed (the default) or
can they be optimized over?}
}
\value{
a matrix with one row for each set of breakpoints.
}
\description{
Return a set of breakpoints for fitting a piecewise survival model
}
