% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_exp.R
\name{get_custom_fss_piecewise_exp}
\alias{get_custom_fss_piecewise_exp}
\title{Construct piecewise exponential distribution function with use with survreg}
\usage{
get_custom_fss_piecewise_exp(fixed_bp, nr, nc)
}
\arguments{
\item{fixed_bp}{Should breakpoints be considered fixed?}

\item{nr}{number of rows in the parameter matrix}

\item{nc}{number of columns in the parameter matrix}
}
\value{
A custom distribution function for use with survreg.  We don't
include a function to determine initial values, which will be
provided in the call to flexsurvreg.
}
\description{
Construct piecewise exponential distribution function with use with survreg
}
