% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_general.R
\name{get_custom_fss_piecewise_multi}
\alias{get_custom_fss_piecewise_multi}
\title{Construct piecewise distribution function with use with survreg}
\usage{
get_custom_fss_piecewise_multi(dists, fixed_bp)
}
\arguments{
\item{dists}{The distributions for the consecutive pieces}

\item{fixed_bp}{Should breakpoints be considered fixed?}
}
\value{
A custom distribution function for use with survreg.  We don't
include a function to determine initial values, which will be
provided in the call to flexsurvreg.
}
\description{
Construct piecewise distribution function with use with survreg
}
