% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{get_state_names.part_surv}
\alias{get_state_names.part_surv}
\title{get_state_names method for part_surv objects}
\usage{
get_state_names.part_surv(x, ...)
}
\arguments{
\item{x}{partitioned survival (\code{part_surv}) object}

\item{...}{additional arguments (ignored)}
}
\value{
state names associated with the object
}
\description{
get_state_names method for part_surv objects
}
