% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_find_best_piecewise_survival_models_multi.R
\name{get_survival_arg_list}
\alias{get_survival_arg_list}
\title{Get a list of arguments for fitting a piecewise survival curve}
\usage{
get_survival_arg_list(these_bp, dists, fixed_bp, survtimes)
}
\arguments{
\item{these_bp}{break points.}

\item{dists}{Sequence of distributions.}

\item{fixed_bp}{Should the break points be considered fixed?}

\item{survtimes}{Survival times for the analysis.}
}
\value{
a list of arguments to be used in
\code{\link[flexsurv]{flexsurvreg}}
}
\description{
Get a list of arguments for fitting a piecewise survival curve
}
