% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit_output.R
\name{plot_fit_data}
\alias{plot_fit_data}
\title{Plot fit data}
\usage{
plot_fit_data(data_to_plot, plot_type = c("survival", "cumulative hazard",
  "log cumulative hazard"), plot_groups = NULL, groups = "dist",
  scale_time = 1, time_label = "time", max_scaled_time = Inf,
  title = NULL, x_axis_gap, legend_loc = "right", logy = NULL,
  km_width = 1.25, label_size = 1.25, silence_warnings = FALSE)
}
\arguments{
\item{data_to_plot}{a data frame from
\code{\link[=prepare_plot_data_from_fit_tibble]{prepare_plot_data_from_fit_tibble()}}}

\item{plot_type}{\code{survival} or \code{cumulative hazard}}

\item{plot_groups}{which groups should be included
in the plot.   If NULL, all distributions in the data
are included.}

\item{groups}{column name to use for grouping.}

\item{scale_time}{times are multiplied by this.  So if your
fit was done on a time scale of days, and you want to plot
by weeks, set this to 1/7.}

\item{time_label}{the \code{x} axis label}

\item{max_scaled_time}{maximum time, after scaling by \code{scale_time}}

\item{title}{for the plot}

\item{x_axis_gap}{distance between breaks on the x axis}

\item{legend_loc}{location for legend - "top", "bottom", "left", or "right"}

\item{logy}{should the \code{y} axis be on the logarithmic scale?  By default,
it is TRUE for survival curves and FALSE for cumulative hazard.}

\item{km_width}{width for the Kaplan-Meier curve line;
approximately 0.5 seems to be the standard width for lines}

\item{label_size}{relative multiplier for label size.}

\item{silence_warnings}{suppresses certain warnings}
}
\value{
a \code{ggplot2} plot object
}
\description{
Plot fit data
}
