% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit_output.R
\name{prepare_plot_data_from_fit_tibble}
\alias{prepare_plot_data_from_fit_tibble}
\title{Prepare fit data for plotting}
\usage{
prepare_plot_data_from_fit_tibble(fit_tib, B_ci = 100)
}
\arguments{
\item{fit_tib}{a tibble containing fits}

\item{B_ci}{number of simulations to use for confidence intervals
for \code{flexsurvreg} objects; becomes the \code{B} of
\code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}
}
\value{
a tibble with the necessary data
}
\description{
Prepare fit data for plotting
}
