% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_general.R
\name{r_piecewise_builder}
\alias{r_piecewise_builder}
\title{Create a random number generator for a piecewise  distribution}
\usage{
r_piecewise_builder(piecewise_def)
}
\arguments{
\item{piecewise_def}{a data frame with columns \code{start} and
\code{dist_name}, and additional columns for parameters of distributions.
\code{dist_name[i]} along with its parameters define the hazard
from \code{start[i]} to
\code{start[i+1]} (or all remaining time in the case of the
last distribution).}
}
\value{
A function that takes in a number \code{n} and returns
\code{n} random draws from the piecewise distribution
defined by \code{piecewise_def}
}
\description{
Create a random number generator for a piecewise  distribution
}
\examples{
my_pe <- data.frame(start = c(0, 200), hazard = c(0.005, 0.025))
r_my_pe <- r_piecewise_exponential_builder(my_pe)
plot(survival::survfit(survival::Surv(r_my_pe(1000), rep(1, 1000)) ~ 1), 
     fun = "cumhaz")
abline(0, 0.005, col = "red")
abline(-200 * (0.025 - 0.005), 0.025, col = "red", lty = 2)
}
