% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_exp.R
\name{r_piecewise_exponential_builder}
\alias{r_piecewise_exponential_builder}
\title{Create a random number generator for a piecewise exponential distribution}
\usage{
r_piecewise_exponential_builder(piecewise_exponential)
}
\arguments{
\item{piecewise_exponential}{a data frame with columns \code{start} and
\code{hazard}.  \code{hazard[i]} is the hazard from \code{start[i]} to
\code{start[i+1]} (or all remaining time in the case of the
last \code{hazard}).}
}
\value{
A function that takes in a number \code{n} and returns
\code{n} random draws from the piecewise exponential distribution
defined by \code{piecewise_exponential}
}
\description{
Create a random number generator for a piecewise exponential distribution
}
\examples{
my_pe <- data.frame(start = c(0, 200), hazard = c(0.005, 0.025))
r_my_pe <- r_piecewise_exponential_builder(my_pe)
plot(survival::survfit(survival::Surv(r_my_pe(1000), rep(1, 1000)) ~ 1), 
     fun = "cumhaz")
abline(0, 0.005, col = "red")
abline(-200 * (0.025 - 0.005), 0.025, col = "red", lty = 2)
}
