% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{subtract_times}
\alias{subtract_times}
\title{Subtract time}
\usage{
subtract_times(subset_data, to_subtract, this_time_col, set_name, condition)
}
\arguments{
\item{subset_data}{some data}

\item{to_subtract}{the number to subtract from times}

\item{this_time_col}{the name of the time column}

\item{set_name}{in case needed for error message}

\item{condition}{in case needed for error message}
}
\value{
the data, with `to_subtract`` subtracted from the time column
}
\description{
Subtract time
}
\details{
This is for fitting subsets that are defined by
being after a certain time, and we want to
fit as if the cutoff time is the new base time.

Moved into a separate function to get the error check
and lengthy construction of an error messageout of the main flow.
}
