% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{transform_tm_info.part_surv_tm_info}
\alias{transform_tm_info.part_surv_tm_info}
\title{Title}
\usage{
transform_tm_info.part_surv_tm_info(tm_info, state_names, ...)
}
\arguments{
\item{tm_info}{specification of a transition object; either a
partial survival object (defined in this package) or a matrix
(defined in  package \code{heemod})}

\item{state_names}{names of the states the model is to be used with}

\item{...}{additional arguments; for partitioned survival objects,
reference file and environment}
}
\value{
a partial survival object
}
\description{
Title
}
