% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{weighted_dose_costs}
\alias{weighted_dose_costs}
\alias{weighted_norm_dose_costs}
\alias{weighted_lognorm_dose_costs}
\title{Compute weighted drug cost over a distribution of characteristics}
\usage{
weighted_dose_costs(dist, params, var_base, dose_base, dose_multiplier,
  available_units, subset_col, subset_val, share_vials, qmin = 0.01,
  qmax = 0.99, by = 0.01)

weighted_norm_dose_costs(mean, sd, var_base, dose_base, dose_multiplier,
  available_units, subset_col, subset_val, share_vials, qmin = 0.01,
  qmax = 0.99, by = 0.01)

weighted_lognorm_dose_costs(meanlog, sdlog, var_base, dose_base,
  dose_multiplier, available_units, subset_col, subset_val, share_vials,
  qmin = 0.01, qmax = 0.99, by = 0.01)
}
\arguments{
\item{dist}{distribution for the dose-determining variable}

\item{params}{arguments for the quantile function for \code{dist}}

\item{var_base}{see details}

\item{dose_base}{see details}

\item{dose_multiplier}{see details}

\item{available_units}{available sizes and costs; must have column names
'size' and 'cost'}

\item{subset_col}{optionally, a subset to select on in available_units}

\item{subset_val}{if \code{subset_col} exists, the value to select}

\item{share_vials}{should vials be shared (or, looked at from the
other side, should any drug be allowed to go unused?)}

\item{qmin, qmax, by}{determine the sequence of quantiles of the variable
at which the dose function will be evaluated}

\item{mean}{mean of normal distribution}

\item{sd}{sd of normal distribution}

\item{meanlog}{mean of lognormal distribution}

\item{sdlog}{sd of lognormal distribution}
}
\value{
the weighted dose cost
}
\description{
Compute weighted drug cost over a distribution of characteristics
}
\details{
Typical use for these functions are when dose varies by some
patient characteristic, for example weight or body surface area.

The general formula is
dose = base_dose + dose_multiplier * pmax(0, var - var_base),
and the number of vials is then computed using
\code{\link[=find_least_cost_partition]{find_least_cost_partition()}}.

Changing the percentiles at which the doses are calculated by
changing \code{qmin}, \code{qmax}, and \code{by} will produce slightly different
values, but the values will converge if

The functions \code{\link[=weighted_norm_dose_costs]{weighted_norm_dose_costs()}} and
\code{\link[=weighted_lognorm_dose_costs]{weighted_lognorm_dose_costs()}} are provided to cover two common cases.
}
\examples{
vialCost <- data.frame(treatment = "fake", 
                      size = c(50, 250), 
                      cost = c(2521, 12297)
                      )
weighted_dose_costs("norm", params = c(mean = 1.85, sd = 0.25),
                    var_base = 0, 
                    dose_base = 0, 
                    dose_multiplier = 320, 
                    vialCost, "treatment", "fake",
                    share_vials = FALSE)
vialCost <- data.frame(treatment = "fake", 
                      size = c(50, 250), 
                      cost = c(2521, 12297)
                      )
weighted_norm_dose_costs(1.85, 0.25, 
                         var_base = 0, dose_base = 0, 
                         dose_multiplier = 320, vialCost, "treatment", "fake",
                         share_vials = FALSE)

vialCost <- data.frame(treatment = "fake", size = 50, cost = 1000)
weighted_lognorm_dose_costs(4.3423559, 0.2116480, 
                            var_base = 50, dose_base = 100, 
                            dose_multiplier = 2, vialCost, "treatment", "fake",
                            share_vials = FALSE)
weighted_lognorm_dose_costs(4.3423559, 0.2116480, 
                            var_base = 50, dose_base = 100, 
                            dose_multiplier = 2, vialCost, "treatment", "fake",
                            share_vials = TRUE)
}
